+++ RELEASES.md
@@ -733,3 +733,4 @@
 11.2.4 2020-03-16
 -----------------
   - refactor: prioritization of data providers by length of filter pattern desc
+  - fix: when using volatile permissions, system permissions starting with _ are retained in same datastore as groups